get rid of safe mod
