+++ src/lib.rs
@@ -2,7 +2,6 @@ pub mod ffi;
 mod odbc_object;
 mod raii;
 mod diagnostics;
-mod safe;
 mod result;
 mod environment;
 mod data_source;
@@ -20,4 +19,21 @@ pub trait Handle{
     type To;
     /// Returns a valid handle to the odbc type.
     unsafe fn handle(&self) -> * mut Self::To;
+}
+
+fn as_out_buffer(buffer: &mut [u8]) -> *mut u8 {
+    if buffer.len() == 0 {
+        std::ptr::null_mut()
+    } else {
+        buffer.as_mut_ptr()
+    }
+}
+
+fn as_buffer_length(n: usize) -> ffi::SQLSMALLINT {
+    use std;
+    if n > std::i16::MAX as usize {
+        std::i16::MAX
+    } else {
+        n as i16
+    }
 }
