+++ src/result.rs
@@ -17,10 +17,7 @@ pub enum Return<T> {
 /// Allocating an environment is the one operation in ODBC which does not yiel a diagnostic record
 /// in case of an error. There is simply no Environment to ask for a diagnostic record
 #[derive(Debug)]
-pub enum EnvAllocError {
-    Info(DiagnosticRecord),
-    Error,
-}
+pub struct EnvAllocError;
 
 impl Display for EnvAllocError {
     fn fmt(&self, f: &mut Formatter) -> std::fmt::Result {
@@ -31,10 +28,7 @@ impl Display for EnvAllocError {
 
 impl std::error::Error for EnvAllocError {
     fn description(&self) -> &str {
-        match self {
-            &EnvAllocError::Info(ref dr) => dr.description(),
-            &EnvAllocError::Error => "Failure to allocate ODBC environment",
-        }
+        "Failure to allocate ODBC environment"
     }
 
     fn cause(&self) -> Option<&std::error::Error> {
